requirements.txt

fastapi==0.112.0
uvicorn==0.30.0
pydantic==2.8.2
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
pytest==8.2.0
pycryptodome==3.20.0
qrcode==7.4.2
